refactor(markdown): extract MarkDownImpl pure helpers and code block
Move the fence metadata, stable-block splitter, repo-path resolution and
the CodeBlock memo into markdownCodeFence / markdownStableBlocks /
markdownRepoPath / MarkdownCodeBlock, and cover the streaming splitter
with tests. The markdownComponents memo stays put; behaviour unchanged.
